Refactor durable activities and orchestrators
