+++ src/Acmebot.App/Extensions/HttpContentExtensions.cs
@@ -4,7 +4,7 @@ namespace Acmebot.App.Extensions;
 
 internal static class HttpContentExtensions
 {
-    public static async ValueTask<T> ReadAsAsync<T>(this HttpContent content)
+    public static async ValueTask<T?> ReadAsAsync<T>(this HttpContent content)
     {
         await using var contentStream = await content.ReadAsStreamAsync();
