+++ src/Acmebot.App/Functions/Http/GetDnsZones.cs
@@ -1,5 +1,4 @@
-﻿using Acmebot.App.Functions.Orchestration;
-using Acmebot.App.Models;
+﻿using Acmebot.App.Models;
 
 using Azure.Functions.Worker.Extensions.HttpApi;
