+++ src/Acmebot.App/Functions/Http/RevokeCertificate.cs
@@ -1,5 +1,4 @@
 ﻿using Acmebot.App.Extensions;
-using Acmebot.App.Functions.Orchestration;
 
 using Azure.Functions.Worker.Extensions.HttpApi;
