--- src/Acmebot.App/Options/GandiOptions.cs
@@ -1,6 +1,0 @@
-﻿namespace Acmebot.App.Options;
-
-public class GandiOptions
-{
-    public string? ApiKey { get; set; }
-}
