+++ src/Acmebot.App/Program.cs
@@ -111,11 +111,6 @@ builder.Services.AddSingleton<IWebhookPayloadBuilder>(provider =>
         return new TeamsPayloadBuilder();
     }
 
-    if (host.EndsWith(".office.com", StringComparison.OrdinalIgnoreCase))
-    {
-        return new LegacyTeamsPayloadBuilder();
-    }
-
     return new GenericPayloadBuilder(options);
 });
 
@@ -134,7 +129,6 @@ builder.Services.AddSingleton<IEnumerable<IDnsProvider>>(provider =>
     dnsProviders.TryAdd(options.Cloudflare, o => new CloudflareProvider(o));
     dnsProviders.TryAdd(options.CustomDns, o => new CustomDnsProvider(o));
     dnsProviders.TryAdd(options.DnsMadeEasy, o => new DnsMadeEasyProvider(o));
-    dnsProviders.TryAdd(options.Gandi, o => new GandiProvider(o));
     dnsProviders.TryAdd(options.GandiLiveDns, o => new GandiLiveDnsProvider(o));
     dnsProviders.TryAdd(options.GoDaddy, o => new GoDaddyProvider(o));
     dnsProviders.TryAdd(options.GoogleDns, o => new GoogleDnsProvider(o));
